internal/jsonld/converter_test.go
package jsonld

import (
	"testing"

	"github.com/chrlesur/json-ld-converter/internal/parser"
	"github.com/chrlesur/json-ld-converter/internal/schema"
	"github.com/stretchr/testify/assert"
)

// Mock du client LLM pour les tests
type mockLLMClient struct{}

func (m *mockLLMClient) Translate(content, sourceLang, targetLang, additionalInstruction string) (string, error) {
	// Simuler une réponse simple pour les tests
	return "Contenu enrichi: " + content, nil
}

func TestConvert(t *testing.T) {
	// Créer une instance de test du convertisseur
	mockVocabulary := schema.NewVocabulary() // Assurez-vous d'avoir une implémentation de test pour le vocabulaire
	mockLLM := &mockLLMClient{}
	converter := NewConverter(mockVocabulary, mockLLM, 1000, "")

	testCases := []struct {
		name           string
		input          string
		expectedOutput map[string]interface{}
		expectedError  error
	}{
		{
			name:  "Conversion simple",
			input: "Ceci est un test",
			expectedOutput: map[string]interface{}{
				"@context": "https://schema.org",
				"@type":    "Thing",
				"text":     "Contenu enrichi: Ceci est un test",
			},
			expectedError: nil,
		},
		// Ajoutez d'autres cas de test ici
	}

	for _, tc := range testCases {
		t.Run(tc.name, func(t *testing.T) {
			segment := &parser.DocumentSegment{Content: tc.input}
			result, err := converter.Convert(segment)

			if tc.expectedError != nil {
				assert.Error(t, err)
				assert.Equal(t, tc.expectedError.Error(), err.Error())
			} else {
				assert.NoError(t, err)
				assert.Equal(t, tc.expectedOutput, result)
			}
		})
	}
}

func TestHandleNestedStructures(t *testing.T) {
	// Test similaire pour handleNestedStructures
	// ...
}

func TestApplyAdditionalInstructions(t *testing.T) {
	// Test pour applyAdditionalInstructions
	// ...
}

func TestTokenLimitHandling(t *testing.T) {
	// Test pour vérifier la gestion des limites de tokens
	// ...
}

func TestIntegration(t *testing.T) {
	// Test d'intégration simulant un flux complet de conversion
	// ...
}
